% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_pyramid.R
\name{labeller_abs}
\alias{labeller_abs}
\title{Label values by absolute value}
\usage{
labeller_abs(...)
}
\arguments{
\item{...}{Extra arguments passed to `format`}
}
\value{
Function that can be used to format values, best used inside `scale_*` or `geom_label`-like functions.
}
\description{
Label values by absolute value
}
