% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percDF.R
\name{percDF}
\alias{percDF}
\title{Creates percentage of a dataframe}
\usage{
percDF(df, age, sex, pop)
}
\arguments{
\item{df}{Name of dataframe}

\item{age}{Age or age group. Write the parameter in quotation marks.}

\item{sex}{Sex or other categorical grouping variable. Write the parameter in quotation marks.}

\item{pop}{Population (in numerical value). Write the parameter in quotation marks.}
}
\value{
The dataframe in long format and in percentage
}
\description{
Creates a dataframe in long format and in percent
}
\examples{
df <- popPyramid::popPER
df <- dplyr::filter(df, Year==2021)
df <- percDF(df, "Age", "Sex", "Population")
}
