% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popPER.R
\docType{data}
\name{popPER}
\alias{popPER}
\title{Peru population (1995-2030)}
\format{
dataframe
\describe{
  \item{Year}{chr Year}
  \item{Sex}{chr Sex}
  \item{Age}{chr Age}
  \item{gAge}{chr Age group}
  \item{Population}{dbl Population}
}
}
\usage{
popPER
}
\description{
Peru population (1995-2030)
}
\keyword{datasets}
