% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_label_pyramid.R, R/geom_pyramid.R
\docType{data}
\name{GeomLabelPyramid}
\alias{GeomLabelPyramid}
\alias{GeomPyramid}
\alias{popPyramid-ggproto}
\title{Base ggproto classes for popPyramid}
\description{
Base ggproto classes for popPyramid
}
\seealso{
[`ggproto()`][ggplot2::ggproto]
}
\keyword{datasets}
\keyword{internal}
